+++ src/app.module.ts
@@ -11,6 +11,8 @@ import { MailModule } from './mail/mail.module';
 import { AccountsModule } from './accounts/accounts.module';
 import { RecordsModule } from './records/records.module';
 import { CategoriesModule } from './categories/categories.module';
+import { ExpensesModule } from './expenses/expenses.module';
+import { IncomesModule } from './incomes/incomes.module';
 import { BudgetsModule } from './budgets/budgets.module';
 import { BudgetHistoryModule } from './budget-history/budget-history.module';
 import config from './config';
@@ -43,6 +45,8 @@ import config from './config';
     AccountsModule,
     RecordsModule,
     CategoriesModule,
+    ExpensesModule,
+    IncomesModule,
     BudgetsModule,
     BudgetHistoryModule,
   ],