+++ src/expenses/expenses.dto.ts
@@ -1,7 +1,6 @@
 import { IsMongoId, IsNotEmpty, IsBoolean } from 'class-validator';
-import { CreateRecordDto } from './records.dto';
+import { CreateRecordDto } from '../records/dtos/records.dto';
 import { PartialType } from '@nestjs/swagger';
-import { CreateExpense } from '../entities/expenses.entity';
 
 export class CreateExpenseDto extends CreateRecordDto {
   @IsBoolean()
@@ -11,8 +10,14 @@ export class CreateExpenseDto extends CreateRecordDto {
 export class UpdateExpenseDto extends PartialType(CreateExpenseDto) {
   @IsMongoId()
   @IsNotEmpty()
-  readonly recordId: CreateExpense;
+  readonly recordId: string;
 
   @IsMongoId()
   readonly userId: string;
+}
+
+export class DeleteExpenseDto {
+  @IsMongoId()
+  @IsNotEmpty()
+  readonly recordId: string;
 }