+++ src/utils/getMonthNumber.ts
@@ -1,0 +1,4 @@
+export const getMonthNumber = (month: string): number => {
+  const monthNumber = new Date(`${month} 1, 2021`).getMonth();
+  return monthNumber;
+};